move EditProfileModal to seperate page
